remove rust-analyze-compat
The history of this build is we added it in some lower level repos that had complex workspace setups (e.g. env and sdk) and we kept breaking rust-analyzer which was destroying our IDE productivity. Things have changed, the sdk repo is not as complex as it was, the env still is. But in any case none of the problems we had would ever happen in this repo.