Fix typo from #237
I didn't adapt the logic right.

This was hard to debug because, in integration tests, it caused the
daemon to run `d.logger.WithError(err).Fatal("could not build
migrations")`, which in turn runs `os.Exit` which ends the test right
away (without flushing logs etc).

I need to think how to improve this. One option is to
override the `ExitFunc` in logrus with an assertion.